+++ examples/ex-stati-card.R
@@ -50,7 +50,7 @@ ui <- fluidPage(
       statiCard(
         "123456 something very very long",
         "Long value text with icon",
-        icon = icon("tachometer"),
+        icon = icon("tachometer-alt"),
         left = TRUE
       ),
       statiCard(